backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-jose[cryptography]==3.3.0
weasyprint==62.1
resend==0.7.0
python-multipart==0.0.9
httpx==0.26.0
pydantic-settings==2.1.0
